docs: add Doxygen comments across all C++ headers and source files
- Add comprehensive Doxygen comments (@brief, @param, @return, and LaTeX formulas \f$ ... \f$) across all files in src/
- Document Tree spatial decomposition components (Tree, NodeI, Node, Leaf, Morton)
- Document Multipole and Local expansion sets in 2D and 3D (MultipoleSet, LocalSet, PowerSet, Irregular)
- Document physics & math entities (Tensor, Particle, math_utils)
- Document acceleration solvers and integrators (SystemUpdateMethod, NaiveUpdate, FMM, QuadratureMethod, SymplecticEuler, RungeKutta4)

+++ src/system_updaters/multipole_utils/power_set2D.cpp
@@ -1,3 +1,8 @@
+/**
+ * @file power_set2D.cpp
+ * @brief Implementation of 2D PowerSet complex power evaluation.
+ */
+
 #include "power_set2D.hpp"
 #include "math_utils/math_utils.hpp"
 
@@ -8,12 +13,10 @@ PowerSet<2>::PowerSet(unsigned int L, const Tensor &r)
     elements.push_back(1.0);
     elements.push_back(r[0]);
     elements.push_back(r[1]);
-    for(int i=3; i<2*L+1; i=i+2){
+    for(size_t i=3; i<2*L+1; i=i+2){
         std::complex<double> complex_multipole = std::complex(elements[i-2], elements[i-1]) * std::complex(r[0], r[1]);
         elements.push_back(complex_multipole.real());
         elements.push_back(complex_multipole.imag());
-        //elements[i] = elements[i-2]*r[0] - elements[i-1]*r[1];
-        //elements[i+1] = elements[i-2]*r[1] + elements[i-1]*r[0];
     } 
 }
